Add controller reference to controlled resources
The `Object` resource provisions other kubernetes resources. However those resulting resources do not have an owner reference set, which leaves them abandoned on the cluster once the `Object` resource gets deleted.

This properly sets the owner reference on the controlled resource using controllerutils.SetControllerReference.

A *runtime.Scheme is needed for the function which can be optained via external.client.Scheme(). Rather than passing that *runtime.Scheme as parameter, I made `getDesired` a receiver function of `external`, so it would have direct access to it. It was a private function anyway, so no breaking changes

Signed-off-by: Jonny Langefeld <jonny.langefeld@getcruise.com>