cleanerr: Clean errors for stable status
Some error messages from the kubernetes API contain formatted pointers. This
triggers an infinite loop in provider-kubernetes, as each attempt to reconcile
updates the status with a new error message.

To avoid this loop, mask pointer references in errors on update.

A similar fix exists in provider-aws.

Without this change, we can see a lot of reconciles for one object over a short
time period:

  Warning  CannotUpdateExternalResource  3m38s (x849 over 8m39s)  managed/object.kubernetes.crossplane.io  (combined from similar events):

This might also block other resources from reconciling.

Signed-off-by: Carl Henrik Lunde <chlunde@ifi.uio.no>