Add `ObservedObjectCollection` API type
Objects in the collection are defined by:
* GVK
* optional namespace
* label selector

The objects are fetched using the specified provider config
and for the matched objects the provider creates counterpart
observe-only objects in the local cluster.

The created objects are owned by the collection resource and
reconciled as usual by the provider. They are also referenced
in collection's `.status.objects` field.

Signed-off-by: Predrag Knezevic <predrag.knezevic@upbound.io>